First version of the struct -> OpenAPI tool
This is a VERY VERY VERY rough first iteration.

The idea is the following:

We have a command line tool that receives a rust file, parses it,
extracts the struct in that file and converts it into an OpenAPI model.

Right now the current implementation is written to just pass the test on
`examples/simple_struct`, but will likely fail miserably with anything a
little bit more complex.

Next things to improve:

- Clean code.
- Add more error handling and less `unwraps`.
- Support all the types and not only the ones tested so far.
- Add the `items` field into the `OpenAPISchema` struct to make `Vec`
compatible.

Example:

```
$ cargo run examples/simple_struct.rs
   Compiling openapi_generator v0.1.0
(/Users/agustinberrini/dev/experiments/cdd-rust/openapi_generator)
    Finished dev [unoptimized + debuginfo] target(s) in 1.52s
     Running `target/debug/openapi_generator examples/simple_struct.rs`
---
swagger: "3.0"
info:
  title: Autogenerated OpenAPI model
  version: "1.0"
paths: []
components:
  - schemas:
      SimpleStruct:
        required:
          - id
          - name
        properties:
          id:
            format: int32
            type: integer
          name:
            format: ~
            type: string
          optional_value:
            format: int64
            type: integer
          list_of_values:
            format: ~
            type: array
```